feat(core): extract plan-validation + process to @karajan/core (KJC-TSK-0511)
PR6 of KJC-TSK-0511. Moves the v2 plan blocked_by validator and the
execa-based runCommand wrapper to @karajan/core so @karajan/hu-board
can replace the last two STATIC ../../../src imports it had.

- packages/core/src/{plan-validation,process}.js: verbatim copies.
- src/plan/plan-validation.js, src/utils/process.js: thin shims.
- packages/hu-board/src/plan-mutations.js + routes/wiki.js: imports
  now resolve via @karajan/core/* subpaths.
- tests/wiki-route.test.js: vi.mock target updated to match.
- 8 new tests in packages/core/tests (cycle detection, ref validity,
  spawn ENOENT enrichment, timeout, stdout capture).

With this PR hu-board has zero static relative imports into the CLI
src tree. The remaining `await import(...)` calls into brain/, rag/,
git/hu-snapshot are dynamic and survive the package split; a follow-up
will extract them next.

Net delta +78 LOC (well under the 200 budget).
