+++ packages/core/src/index.js
@@ -11,3 +11,5 @@ export * from "./run-registry.js";
 export * from "./vec-store.js";
 export * from "./plan-id.js";
 export * from "./plan-hu-ops.js";
+export * from "./plan-validation.js";
+export * from "./process.js";
