+++ packages/hu-board/src/plan-mutations.js
@@ -23,7 +23,7 @@ import { trackRun, untrack } from './run-tracker.js';
 import { getHuBoardRunsDir } from './db.js';
 import { fileURLToPath } from 'node:url';
 import { updateHuStatus, certifyAllHus, updateHu } from '@karajan/core/plan-hu-ops';
-import { validateBlockedByChange } from '../../../src/plan/plan-validation.js';
+import { validateBlockedByChange } from '@karajan/core/plan-validation';
 import { syncPlanFile } from './sync.js';
 import { getDb } from './db.js';
